Started to fix the execution of the nodes.
