+++ hermes/Resources/executers/jinjaExecuters.py
@@ -7,30 +7,6 @@ from jinja2 import FileSystemLoader, Environment
 from hermes.Resources.executers.abstractExecuter import abstractExecuter
 
 # *************************************************************
-class jinjaExecuter(abstractExecuter):
-
-    def _defaultParameters(self):
-        return dict(
-            output=["status"],
-            inputs=["classpath", "function"],
-            webGUI=dict(JSONSchema="webGUI/jinjaExecuter_JSONchema.json",
-                        UISchema="webGUI/jinjaExecuter_UISchema.json"),
-            parameters={}
-        )
-
-    def _getTemplate(self,templateName,additionalTemplatePath=[]):
-
-        templatePath = [os.path.join(pathlib.Path(__file__).parent.absolute(), "jinjaTemplates")] + list(additionalTemplatePath)
-
-        tmp_path = os.path.join(pathlib.Path(__file__).parent.absolute(), "jinjaTemplates")
-        print("tmp_path = " + str (tmp_path) + "\n")
-        print("additionalTemplatePath = " + str(additionalTemplatePath) + "\n")
-        print("templatePath = " + str (templatePath) + "\n")
-        print("templateName = " + templateName + "\n")
-
-        file_loader = FileSystemLoader(templatePath)
-        env = Environment(loader=file_loader)
-        return env.get_template(templateName)
 
     def run(self, **inputs):
         # get the  name of the template
